gpu provision fe
